refactor: refined all std and keywords
- feat/refactor/fix: stdlib expansion, language features, runtime cleanup
- Add `defer` with closure capture and panic unwinding; introduce explicit `mut`
- Major stdlib reorg: `std.collections→std.core`, `std.strings→std.str`, `fs→std.os.fs`, `itertools→std.core.iter`, I/O→`std.str.io`; remove `std.primitives`, deprecated tests, and `gmtime`
- Consolidate generic collection ops into `std.core.reflect`; add core structures (list, dict, set, queue, heap, sort), networking (HTTP, URL), and expanded utils (json, msgpack, base64, regex, glob, path, fmt, timefmt)
- Add `os()`/`arch()` to `std.os.mod`; fix system-wide std loading
- Improve parser and diagnostics: better errors, fuzzy undefined-symbol suggestions, string unescaping, `normalize_module_name`
- Refactor runtime memory: arena allocation for AST/codegen, centralized teardown, custom `__malloc/__free` with ASan/LSan/gprof
- Tidy std usage and tests for `mut`; refactor webdoc (new UI, better parsing, CLI)
- Expand REPL with multi-line editing and history navigation
- Update docs (`README.md`, `LANGUAGE.md`); remove `METAPROGRAMMING.md`

+++ src/rt/math.c
@@ -111,7 +111,7 @@ FLT_CMP(eq, ==)
 // Mixed Math
 int64_t __add(int64_t a, int64_t b) {
   if (is_int(a) && is_int(b))
-    return a + b - 1;
+    return (int64_t)((uint64_t)a + (uint64_t)b - 1);
   if (is_v_flt(a) || is_v_flt(b))
     return __flt_add(a, b);
   if (is_any_ptr(a) && is_int(b))
@@ -125,7 +125,7 @@ int64_t __add(int64_t a, int64_t b) {
 
 int64_t __sub(int64_t a, int64_t b) {
   if (is_int(a) && is_int(b))
-    return a - b + 1;
+    return (int64_t)((uint64_t)a - (uint64_t)b + 1);
   if (is_v_flt(a) || is_v_flt(b))
     return __flt_sub(a, b);
   if (is_any_ptr(a) && is_int(b))
